requirements.txt
# Что: фреймворк Telegram-бота на aiogram 3.
# Зачем: современная асинхронная архитектура хендлеров и роутеров.
aiogram==3.13.1
# Что: загрузка переменных окружения из .env.
# Зачем: безопасно хранить токен вне кода.
python-dotenv==1.0.1
